pyproject.toml
[project]
name = "alliance-docs-mcp"
version = "0.1.0"
description = "MCP server for Alliance documentation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.12.5",
    "pypandoc>=1.15",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "rich>=14.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/alliance_docs_mcp"]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "pytest>=8.4.2",
    "ruff>=0.14.1",
]
